requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
dominate==2.6.0
Flask==2.0.3
Flask-Bootstrap4==4.0.2
Flask-WTF==1.0.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
requests==2.27.1
typing-extensions==4.1.1
urllib3==1.26.9
visitor==0.1.3
Werkzeug==2.0.3
WTForms==3.0.1
zipp==3.7.0
aiohttp~=3.8.1